+++ Sora/PackageInfo.swift
@@ -9,16 +9,16 @@ public enum SDKInfo {
  */
 public enum WebRTCInfo {
     /// WebRTC フレームワークのバージョン
-    public static let version = "M119"
+    public static let version = "M121"
 
     /// WebRTC フレームワークのコミットポジション
-    public static let commitPosition = "2"
+    public static let commitPosition = "4"
 
     /// WebRTC フレームワークのメンテナンスバージョン
-    public static let maintenanceVersion = "1"
+    public static let maintenanceVersion = "0"
 
     /// WebRTC フレームワークのソースコードのリビジョン
-    public static let revision = "52bc9f7c1205f4b731ea0289b059f7d240c1e228"
+    public static let revision = "0f741da200c064aea70a790d2fbf678e930bff39"
 
     /// WebRTC フレームワークのソースコードのリビジョン (短縮版)
     public static var shortRevision: String {