+++ CHANGES.md
@@ -24,7 +24,7 @@
   - @miosakuma
 - [ADD] VideoCodec に H265 を追加する
   - @enm10k
-- [UPDATE] WebRTCConfiguration に degradationPreference を追加する
+- [ADD] WebRTCConfiguration に degradationPreference を追加する
   - @enm10k
 - [UPDATE] 解像度に qHD (960x540) を追加する
   - @enm10k